+++ src/main/java/ro/itschool/repository/LiteratureRepository.java
@@ -23,7 +23,7 @@ public interface LiteratureRepository extends JpaRepository<Literature, UUID> {
 
     @Query(value = "SELECT * FROM literature WHERE user_id = ?",
             nativeQuery = true)
-    List<Literature> findByUserId(Integer userId);
+    List<Literature> findByUserId(Long userId);
 
     @Query(value = "SELECT * FROM literature l WHERE l.name LIKE %:keyword% OR l.author LIKE %:keyword% OR l.location LIKE %:keyword% OR l.movement LIKE %:keyword% " +
                     "OR l.is_temporary LIKE %:keyword% OR l.id LIKE %:keyword% OR l.user_id LIKE %:keyword% OR l.year LIKE %:keyword%",
