+++ src/main/java/ro/itschool/repository/MusicRepository.java
@@ -23,7 +23,7 @@ public interface MusicRepository extends JpaRepository<Music, UUID> {
 
     @Query(value = "SELECT * FROM music WHERE user_id = ?",
             nativeQuery = true)
-    List<Music> findByUserId(Integer userId);
+    List<Music> findByUserId(Long userId);
 
     @Query(value = "SELECT * FROM music m WHERE m.name LIKE %:keyword% OR m.location LIKE %:keyword% OR m.movement LIKE %:keyword% " +
                     "OR m.is_temporary LIKE %:keyword% OR m.id LIKE %:keyword% OR m.user_id LIKE %:keyword% OR m.event_time LIKE %:keyword%",
