+++ src/main/java/ro/itschool/repository/SculptureRepository.java
@@ -23,7 +23,7 @@ public interface SculptureRepository extends JpaRepository<Sculpture, UUID> {
 
     @Query(value = "SELECT * FROM sculpture WHERE user_id = ?",
             nativeQuery = true)
-    List<Sculpture> findByUserId(Integer userId);
+    List<Sculpture> findByUserId(Long userId);
 
     @Query(value = "SELECT * FROM sculpture s WHERE s.name LIKE %:keyword% OR s.author LIKE %:keyword% OR s.location LIKE %:keyword% OR s.movement LIKE %:keyword% " +
                     "OR s.is_temporary LIKE %:keyword% OR s.id LIKE %:keyword% OR s.user_id LIKE %:keyword% OR s.year LIKE %:keyword%",
